split user tests into a new package
